Move types.Header and friends into their own file
This file and gen_header_json.go will move out soon.
